requirements.txt
beautifulsoup4==4.11.1
google_api_python_client==2.86.0
mysql_connector_repackaged==0.3.1
Pillow==9.2.0
Pillow==9.5.0
protobuf==3.20.1
reportlab==3.6.13
selenium==4.9.1
twilio==8.1.0
webdriver_manager==3.8.6
